# Optimize KnownColor value lookup via a combined color table