--- a/src/libraries/System.Drawing.Primitives/src/System/Drawing/Color.cs
+++ b/src/libraries/System.Drawing.Primitives/src/System/Drawing/Color.cs
@@ -374,7 +374,7 @@ namespace System.Drawing
         public bool IsSystemColor => IsKnownColor && IsKnownColorSystem((KnownColor)knownColor);
 
         internal static bool IsKnownColorSystem(KnownColor knownColor)
-            => KnownColorTable.s_colorTable[(int)knownColor, 0] == (uint)KnownColorTable.KnownColorCategory.System;
+            => KnownColorTable.s_colorTable[(int)knownColor, 0] == KnownColorTable.KnownColorCategorySystem;
 
         // Used for the [DebuggerDisplay]. Inlining in the attribute is possible, but
         // against best practices as the current project language parses the string with