* Fixes: #49161 related to commit d21fe17023e2a03f9b
* Brings the following dotnet/performance microbenchmarks back on par
  with .NET5
  * System.Drawing.Tests.Perf_Color.GetBrightness
  * System.Drawing.Tests.Perf_Color.GetSaturation
  * System.Drawing.Tests.Perf_Color.GetHue
* Use one known color table for both system and non-system
  colors (indexed by the corresponding KnownColor enum value).
* This table is split into 2 arrays: color value and color kind (category).
* Add color category constants for Unknown, System, and Web.
* Combined table removes the need to handle mismatches corresponding
  with the KnownColor enum value.
* Obtaining the value of and verifying if a known color is a system
  color is simplified to a single lookup instead of multiple base
  bounds comparisons inside the fragmented KnownColor enum table.
* These base bounds comparisons would grow linearly as the KnownColor
  enum table had more colors added to it, mixed between system and
  non-system. Now no comparisons are needed.