@L2 2021-03-31
dotnet/performance microbenchmarks filtered with System.Drawing.Tests.Perf_Color*:

Base = Before this changeset
Diff = After this changeset

No Slower results for the provided threshold = 2% and noise filter = 25ns.

| Faster                                        | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Drawing.Tests.Perf_Color.GetBrightness |      1.48 |          1540.92 |          1038.91 |         |
| System.Drawing.Tests.Perf_Color.GetSaturation |      1.47 |          1622.30 |          1104.48 |         |
| System.Drawing.Tests.Perf_Color.GetHue        |      1.35 |          1755.29 |          1303.64 |         |