@L2 2021-04-02
What my profiling showed was that Color.IsKnownColorSystem(color) called here was not being inlined anymore due to the extra comparison that was added to the KnownColorToArgb(KnownColor color) method:

https://github.com/dotnet/runtime/blob/5078dd61a81c88dbdd131a77b8148b4dee3803fa/src/libraries/Common/src/System/Drawing/KnownColorTable.cs#L187-L199

I also saw that there was a lot of time being spent in IsKnownColorSystem(KnownColor knownColor) due to the extra comparisons that were added:
https://github.com/dotnet/runtime/blob/5078dd61a81c88dbdd131a77b8148b4dee3803fa/src/libraries/System.Drawing.Primitives/src/System/Drawing/Color.cs#L375-L376

To mitigate both causes, I combined the color lookup tables inside KnownColorTable.cs, indexed by their enum value in KnownColor.cs. I also augmented this table to include the color category which leads to no base bounds comparisons necessary to check if a color is categorized as system or non-system (web).

Regarding the base bounds comparisons: the existing implementation categorized the color based on its location inside the KnownColor enum table. As this table became more fragmented, more comparisons would be needed to categorize a color. 

For example:
Enum {
SystemColor1
SystemColor2
SystemColor3
Web1
Web2
SystemColor4
SystemColor5
SystemColor6
Web3
SystemColor7
Web4}

Since modifications to the KnownColor enum table are breaking changes, new colors are added to the end and the existing implementation checks if a given known color is a system color by checking if its inside the enum value bounds of (SystemColor1 to SystemColor3 | SystemColor4 to SystemColor6| SystemColor7) in this example. As this becomes more fragmented, the comparisons grow linearly.

As we already have a unique lookup value for each color (their enum value) we can use a secondary table as it already exists in two parts (one for web and another for system) in KnownColorTable, combine it, and also include the category labels.