@DrewScoggins 2021-04-14
### Improvemnts in System.Drawing.Tests.Perf_Color

Benchmark | Baseline | Test | Test/Base | Baseline IR | Compare IR | IR Ratio | Baseline ETL | Compare ETL
-- | -- | -- | -- | -- | -- | -- | -- | --
[GetHue](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Drawing.Tests.Perf_Color.GetHue.html>) | 1.65 μs | 1.36 μs | 0.83 | | |
[GetSaturation](<https://pvscmdupload.blob.core.windows.net/reports/allTestHistory/refs/heads/main_x64_ubuntu 18.04/System.Drawing.Tests.Perf_Color.GetSaturation.html>) | 1.39 μs | 1.15 μs | 0.82 | | |

![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_13_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Drawing.Tests.Perf_Color_1.png>)
![graph](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_13_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Drawing.Tests.Perf_Color_9.png>)
![graph]
[Historical Data in Reporting System](<https://pvscmdupload.blob.core.windows.net/autofilereport/autofilereports/04_13_2021/refs/heads/main_x64_ubuntu%2018.04_Improvement/System.Drawing.Tests.Perf_Color.html>)

We are seeing these results in the lab, thanks for tracking down and fixing this issue.